USBDevice: do not use automatic multi-packet USB transactions
If the host sends an unfinished OUT transaction (=> a transaction
that ends with a data packet filled to the maximum allowed size but
not followed by another shorter packet or a ZLP) the Transaction
Completed interrupt is not fired and the incoming data remains
suspendend until the host sends more data.

This commit fix the problem by disabilng multi-packet transactions.

upstream commit 3316756b0ce32d334b8798206c5bb88984b396b1.

Signed-off-by: Mathias Tillman <master.homer@gmail.com>